#153 created focal point lut.
